Fix bridge-app init
- Bridge app did not have proper CommissionableDataProvider
  init and nobody caught it until some testing weeks later.
- This PR adds the necessary init calls and also refactors
  elements of the bridge app structure to prevent this from
  happening again.

Fixes #15134

Testing done:
- All unit tests pass
- Cert tests pass
- chip-bridge-app works and is commissionablen by chip-tool and chip-repl
