+++ examples/bridge-app/linux/BUILD.gn
@@ -22,13 +22,13 @@ executable("chip-bridge-app") {
   sources = [
     "${chip_root}/examples/tv-app/tv-common/include/CHIPProjectAppConfig.h",
     "Device.cpp",
-    "Options.cpp",
     "include/Device.h",
     "main.cpp",
   ]
 
   deps = [
     "${chip_root}/examples/bridge-app/bridge-common",
+    "${chip_root}/examples/platform/linux:app-main",
     "${chip_root}/src/lib",
   ]
